# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProConnect professional networking and remote jobs platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/{profileId}/portfolio:
    get:
      operationId: listPortfolio
      tags: [portfolio]
      summary: List portfolio projects for a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of portfolio projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioProject"
    post:
      operationId: createPortfolioProject
      tags: [portfolio]
      summary: Add portfolio project
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioProjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioProject"
  /profiles/{profileId}/portfolio/{id}:
    put:
      operationId: updatePortfolioProject
      tags: [portfolio]
      summary: Update portfolio project
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePortfolioProjectBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioProject"
    delete:
      operationId: deletePortfolioProject
      tags: [portfolio]
      summary: Delete portfolio project
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /profiles/{profileId}/portfolio/reorder:
    put:
      operationId: reorderPortfolio
      tags: [portfolio]
      summary: Reorder portfolio projects
      parameters:
        - name: profileId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items: { type: integer }
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
  /profiles/{profileId}/portfolio/import:
    post:
      operationId: importPortfolio
      tags: [portfolio]
      summary: Confirm portfolio import candidates
      parameters:
        - name: profileId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items: { $ref: "#/components/schemas/CreatePortfolioProjectBody" }
      responses:
        "201":
          description: Imported projects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PortfolioProject" }
  /profiles/{profileId}/portfolio/import/github:
    get:
      operationId: discoverGithubPortfolio
      tags: [portfolio]
      summary: Discover public GitHub repositories for import
      parameters:
        - name: profileId
          in: path
          required: true
          schema: { type: integer }
        - name: username
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Repository candidates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GithubPortfolioCandidate" }
...
components:
  schemas:
...
    PortfolioProject:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        projectUrl:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        source:
          type: string
        externalId:
          type: string
          nullable: true
        canonicalUrl:
          type: string
          nullable: true
        objectPath:
          type: string
          nullable: true
        mimeType:
          type: string
          nullable: true
        fileSize:
          type: integer
          nullable: true
        visibility:
          type: string
          enum: [public, private]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, profileId, title, tags, featured, source, visibility, sortOrder, createdAt, updatedAt]
    GithubPortfolioCandidate:
      type: object
      properties:
        externalId: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        canonicalUrl: { type: string, format: uri }
        projectUrl: { type: string, format: uri }
        imageUrl: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        source: { type: string, enum: [github] }
      required: [externalId, title, canonicalUrl, projectUrl, tags, source]
    CreatePortfolioProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        projectUrl:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        source:
          type: string
          enum: [manual, github, behance, dribbble, linkedin, personal, upload]
        externalId:
          type: string
          nullable: true
        canonicalUrl:
          type: string
          format: uri
          nullable: true
        visibility:
          type: string
          enum: [public, private]
        objectPath:
          type: string
          nullable: true
        mimeType:
          type: string
          enum: [application/pdf, image/png, image/jpeg, image/webp]
          nullable: true
        fileSize:
          type: integer
          maximum: 15728640
          nullable: true
      required: [title, tags]
    UpdatePortfolioProjectBody:
      type: object
      description: Partial portfolio project update; all properties are optional.
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        projectUrl:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        source:
          type: string
          enum: [manual, github, behance, dribbble, linkedin, personal, upload]
        externalId:
          type: string
          nullable: true
        canonicalUrl:
          type: string
          format: uri
          nullable: true
        visibility:
          type: string
          enum: [public, private]
        objectPath:
          type: string
          nullable: true
        mimeType:
          type: string
          enum: [application/pdf, image/png, image/jpeg, image/webp]
          nullable: true
        fileSize:
          type: integer
          maximum: 15728640
          nullable: true
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        HMAC-signed token in the format "profileId:hmac-sha256(profileId, SESSION_SECRET)".
        Returned from /auth/login and /auth/register as authToken.